moved controller stuff out of teacup_view_controller, into controller.  moved some layout and view stuff.  DONE!?
