+++ lib/teacup/layout.rb
@@ -139,49 +139,6 @@ module Teacup
       instance
     end
 
-    # Returns a stylesheet to use to style the contents of this controller's
-    # view.  You can also assign a stylesheet to {stylesheet=}, which will in
-    # turn call {restyle!}.
-    #
-    # This method will be queried each time {restyle!} is called, and also
-    # implicitly whenever Teacup needs to draw your layout (currently only at
-    # view load time).
-    #
-    # @return Teacup::Stylesheet
-    #
-    # @example
-    #
-    #  def stylesheet
-    #    if [UIDeviceOrientationLandscapeLeft,
-    #        UIDeviceOrientationLandscapeRight].include?(UIDevice.currentDevice.orientation)
-    #      Teacup::Stylesheet[:ipad]
-    #    else
-    #      Teacup::Stylesheet[:ipadvertical]
-    #    end
-    #  end
-    def stylesheet
-      @stylesheet
-    end
-
-    # Assigning a new stylesheet triggers {restyle!}, so do this during a
-    # rotation to get your different layouts applied.
-    #
-    # Assigning a stylesheet is an *alternative* to returning a Stylesheet in
-    # the {stylesheet} method. Note that {restyle!} calls {stylesheet}, so while
-    # assigning a stylesheet will trigger {restyle!}, your stylesheet will not
-    # be picked up if you don't return it in a custom stylesheet method.
-    #
-    # @return Teacup::Stylesheet
-    #
-    # @example
-    #
-    #   stylesheet = Teacup::Stylesheet[:ipadhorizontal]
-    def stylesheet=(new_stylesheet)
-      @stylesheet = new_stylesheet
-      view.stylesheet = new_stylesheet
-      view.restyle!
-    end
-
     protected
 
     # Get's the current stack of views in nested calls to layout.
